src/requirements.txt
kedro[pandas.CSVDataSet,pandas.ExcelDataSet,pandas.ParquetDataSet,plotly.PlotlyDataSet,plotly.JSONDataSet,spark.SparkDataSet]~=0.18.4        
kedro~=0.18.4
plotnine~=0.9.0
sqlalchemy~=1.4.41
shapely~=2.0.0
numpy == 1.22.4
googlemaps~=4.5.3
polyline~=1.4.0
flexpolyline~=0.1.0
bs4~=0.0.1
pyproj~=3.3.0
PyYAML~=5.3.1
geopandas~=0.10
pandas==1.5.2
pytest==6.2.5
pytest-cov==3.0.0
pytest-mock==1.13.0
anyio==3.6.2
requests-mock==1.10.0
pyspark==3.3.2
xlsxwriter==3.1.0
requires.io==0.2.6
mlflow==2.5.0
cloudpickle==2.0.0
configparser==5.2.0
psutil==5.9.0
typing-extensions~=4.6.0
xgboost==1.4.2
threadpoolctl==3.1.0
azure-storage-blob~=12.23.1
setuptools==70.3.0
